+++ src/tracee/util.cpp
@@ -153,7 +153,7 @@ static constexpr std::array<std::pair<u8, u16>, 24> UserRegsMapping = { { { 8, o
   { 4, offsetof(user_regs_struct, gs) } } };
 #endif
 u64
-get_register(user_regs_struct *regs, int reg_number) noexcept
+get_register(user_regs_struct *regs, size_t reg_number) noexcept
 {
   MDB_ASSERT(reg_number <= 16, "Register number {} not supported", reg_number);
   return *(u64 *)(((std::uintptr_t)regs) + RegNumToX86_64Offsets[reg_number]);
